Harden lifecycle, async auth and null semantics
- Guard in-flight poll/sync callbacks with a lifecycle generation so
  disposed or superseded sessions cannot publish stale state
- Track one-shot channel syncs and cancel them on dispose
- Rebuild the account poller when the API client or device changes so
  reconnects never reuse an obsolete token context
- Apply thing configuration changes at runtime via thingUpdated
- Replace blocking token acquisition on scheduler threads with async
  composition (TokenManager.getAccessTokenAsync)
- Report missing volume/pillow levels as absent instead of zero and
  skip publishing unknown pillow channels
- Deduplicate imports and stale Javadoc in alarm and pillow DTOs

+++ org.openhab.binding.eightsleep/src/test/java/org/openhab/binding/eightsleep/internal/api/EndpointContractTest.java
@@ -350,7 +350,7 @@ public class EndpointContractTest {
         PillowEntry pillow = data.findPillow("left");
         assertNotNull(pillow);
         assertTrue(pillow.isOn());
-        assertEquals(-15, pillow.getLevel());
+        assertEquals(-15, pillow.getLevel().intValue());
         assertTrue(data.containsPod("dev1"));
     }
 
@@ -428,7 +428,7 @@ public class EndpointContractTest {
         }
         assertTrue(state.isPlaying());
         assertFalse(state.isPaused());
-        assertEquals(30, state.getVolumePercent());
+        assertEquals(30, state.getVolumePercent().intValue());
         assertTrue(state.hasSpeaker());
         assertEquals("Rain", state.currentTrack.name);
     }
